requirements.txt
# Neshama Agent 开源版依赖

# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# JWT Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Cache
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# AI Model Adapters
openai>=1.0,<2.0
anthropic>=0.18,<1.0
httpx>=0.25,<1.0

# Vector Store
chromadb>=0.4,<1.0
sentence-transformers>=2.2,<3.0

# File Processing
Pillow>=10.0,<11.0
python-magic>=0.4,<1.0

# Utils
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0

# Development
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0

# Production (optional)
gunicorn>=21.0,<22.0
